Fixes and tests for sampling header
