+++ features/nested_spans.feature
@@ -9,6 +9,7 @@ Feature: Nested Spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:2"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "span1"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.1.name" equals "span2"
 
@@ -26,6 +27,7 @@ Feature: Nested Spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:2"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "span1"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.1.name" equals "span2"
 
@@ -42,6 +44,7 @@ Feature: Nested Spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:3"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "span3"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.1.name" equals "span2"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.2.name" equals "span1"
@@ -66,6 +69,7 @@ Scenario: Make Current Context
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:2"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.parentSpanId" is null
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.1.parentSpanId" is null
