+++ features/scene_load_spans.feature
@@ -9,6 +9,7 @@ Feature: Scene Load Spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:1"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "[ViewLoad/Scene]Scene1"
 
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0" string attribute "bugsnag.span_category" equals "view_load"
@@ -27,6 +28,7 @@ Feature: Scene Load Spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:1"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "[ViewLoad/Scene]Scene1"
 
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0" string attribute "bugsnag.span_category" equals "view_load"
@@ -40,10 +42,11 @@ Feature: Scene Load Spans
 
 Scenario: Load Scene Async
     When I run the game in the "SceneLoadAsync" state
-    And I wait for 1 span
+    And I wait for 3 spans
     Then the trace Bugsnag-Integrity header is valid
     And the trace "Bugsnag-Api-Key" header equals "a35a2a72bd230ac0aa0f52715bbdc6aa"
 
+    * the trace "Bugsnag-Span-Sampling" header equals "1:3"
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0.name" equals "[ViewLoad/Scene]Scene1"
 
     * the trace payload field "resourceSpans.0.scopeSpans.0.spans.0" string attribute "bugsnag.span_category" equals "view_load"
